tests/requirements-dev.txt
pytest==8.3.2
requests==2.32.3
black==24.8.0
httpx==0.27.2
fastapi==0.111.0
uvicorn[standard]==0.30.1
PyJWT==2.9.0
pydantic==2.8.2
pytest-asyncio==1.2.0
pytest-cov==7.0.0